requirements.txt
tensorflow==2.10.0
numpy==1.26.3
Pillow==10.2.0
requests==2.31.0